Commit 0.0.2 - Added DTO's + Mapping functionality - 16:05 - 23.01.2024

+++ src/main/java/com/jrs296/ems/repository/EmployeeRepository.java
@@ -1,7 +1,7 @@
 package com.jrs296.ems.repository;
 
-import com.jrs296.ems.entity.Employee;
+import com.jrs296.ems.models.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
 }
